add crude shiny apps support for #67
AFAICT it's difficult/impossible to dynamically create an nginx reverse
proxy at runtime for each discovered app.

That said, for the purposes of this demo we only need one reverse proxy
to get to the shiny server's radiant:3838/miracle instead of
radiant:3838/base and then construct the appropriate URL to get at the
shiny app / r markdown docs.

In the future we could spin up a shiny server on demand to serve the
request and establish a connection to that new server for some duration,
especially since shiny appears to get overloaded with concurrent
requests. This would likely fall in year 1-2 however as we devote more
time to scaling.

[skip ci]